+++ src/Entity/EntityInterface.php
@@ -6,12 +6,7 @@ namespace Ifrost\DoctrineApiBundle\Entity;
 
 use Ifrost\ApiFoundation\Entity\ApiEntityInterface;
 
-interface EntityInterface extends ApiEntityInterface
+interface EntityInterface extends ApiEntityInterface, WithDbalWritableFormat
 {
-    /**
-     * @return array<string, string|int|bool|float|null>
-     */
-    public function getWritableFormat(): array;
-
     public static function getTableName(): string;
 }