+++ src/Utility/DbClientInterface.php
@@ -4,36 +4,20 @@ declare(strict_types=1);
 
 namespace Ifrost\DoctrineApiBundle\Utility;
 
-use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Exception;
 use Doctrine\DBAL\Types\Type;
 use Ifrost\DoctrineApiBundle\Exception\NotFoundException;
 use Ifrost\DoctrineApiBundle\Query\DbalQuery;
 
-class DbClient implements DbClientInterface
+interface DbClientInterface
 {
-    public function __construct(private Connection $connection)
-    {
-    }
-
     /**
      * Returns the first value of the next row of the result or FALSE if there are no more rows.
      *
      * @throws NotFoundException
      * @throws Exception
      */
-    public function fetchColumn(string|DbalQuery $query, mixed ...$params): mixed
-    {
-        $query = $this->getDbalQuery($query, ...$params);
-        $result = $query->executeQuery();
-        $result = $result->fetchOne();
-
-        if (false === $result) {
-            throw new NotFoundException(sprintf('Record not found for query "%s"', $query::class), 404);
-        }
-
-        return $result;
-    }
+    public function fetchColumn(string|DbalQuery $query, mixed ...$params): mixed;
 
     /**
      * Returns an array containing the values of the first column of the result.
@@ -42,10 +26,7 @@ class DbClient implements DbClientInterface
      *
      * @throws Exception
      */
-    public function fetchFirstColumn(string|DbalQuery $query, mixed ...$params): array
-    {
-        return $this->getDbalQuery($query, ...$params)->executeQuery()->fetchFirstColumn();
-    }
+    public function fetchFirstColumn(string|DbalQuery $query, mixed ...$params): array;
 
     /**
      * Returns the next row of the result as an associative array or FALSE if there are no more rows.
@@ -55,17 +36,7 @@ class DbClient implements DbClientInterface
      * @throws NotFoundException
      * @throws Exception
      */
-    public function fetchOne(string|DbalQuery $query, mixed ...$params): array
-    {
-        $query = $this->getDbalQuery($query, ...$params);
-        $result = $query->executeQuery()->fetchAssociative();
-
-        if (false === $result) {
-            throw new NotFoundException(sprintf('Record not found for query "%s"', $query::class), 404);
-        }
-
-        return $result;
-    }
+    public function fetchOne(string|DbalQuery $query, mixed ...$params): array;
 
     /**
      * Returns an array containing all of the result rows represented as associative arrays.
@@ -74,10 +45,7 @@ class DbClient implements DbClientInterface
      *
      * @throws Exception
      */
-    public function fetchAll(string|DbalQuery $query, mixed ...$params): array
-    {
-        return $this->getDbalQuery($query, ...$params)->executeQuery()->fetchAllAssociative();
-    }
+    public function fetchAll(string|DbalQuery $query, mixed ...$params): array;
 
     /**
      * Inserts a table row with specified data.
@@ -89,10 +57,7 @@ class DbClient implements DbClientInterface
      *
      * @throws Exception
      */
-    public function insert(string $table, array $data, array $types = []): void
-    {
-        $this->connection->insert($table, $data, $types);
-    }
+    public function insert(string $table, array $data, array $types = []): void;
 
     /**
      * Executes an SQL UPDATE statement on a table.
@@ -105,10 +70,7 @@ class DbClient implements DbClientInterface
      *
      * @throws Exception
      */
-    public function update(string $table, array $data, array $criteria, array $types = []): void
-    {
-        $this->connection->update($table, $data, $criteria, $types);
-    }
+    public function update(string $table, array $data, array $criteria, array $types = []): void;
 
     /**
      * Executes an SQL DELETE statement on a table.
@@ -120,24 +82,5 @@ class DbClient implements DbClientInterface
      *
      * @throws Exception
      */
-    public function delete(string $table, array $criteria, array $types = []): void
-    {
-        $this->connection->delete($table, $criteria, $types);
-    }
-
-    private function getDbalQuery(string|DbalQuery $query, mixed ...$params): DbalQuery
-    {
-        if (is_string($query)) {
-            $query = new $query(
-                $this->connection,
-                ...$params
-            );
-        }
-
-        if (!$query instanceof DbalQuery) {
-            throw new \RuntimeException(sprintf('Query is invalid. Query should be instance of %s (%s given).', DbalQuery::class, gettype($query)));
-        }
-
-        return $query;
-    }
+    public function delete(string $table, array $criteria, array $types = []): void;
 }